Serialise the picmi Simulation into setup_dir/metadata (TT-06)
Emit the picmi-level Simulation as metadata/picmi_simulation.json in
Simulation.write_input_file, next to the pypicongpu-level metadata files,
so the user-level input can be reconstructed from a generated setup.

- Simulation.write_input_file stores model_dump(mode="json") after
  generate() and re-snapshots the ro-crate so the new file stays tracked.
- Simulation.from_setup(setup_dir) reloads the picmi Simulation from a
  setup directory via model_validate.
- Make Simulation._post_init tolerant to a raw-dict solver during
  model_validate (skip the cfl/delta_t recompute; both are stored).
- Make picmi.diagnostics.TimeStepSpec serialisable by implementing
  __get_pydantic_core_schema__ so diagnostics using TimeStepSpec[::n]
  no longer break model_dump and round-trip on load-back.
- picongpu_custom_user_input stays excluded from the dump (TT-04/#33);
  the omission is documented. The flattened input remains in
  metadata/pypicongpu_runner.json.
- picongpu_cell_size is annotated as float to stop pydantic serializer
  warnings from the computed field tripping filterwarnings=error.

Closes #35

+++ lib/python/picongpu/picmi/grid.py
@@ -62,7 +62,7 @@ class Cartesian3DGrid(picmistandard.PICMI_Cartesian3DGrid):
     picongpu_super_cell_size: tuple[int, int, int] = Field(default=(8, 8, 4))
 
     @computed_field
-    def picongpu_cell_size(self) -> tuple[int, int, int]:
+    def picongpu_cell_size(self) -> tuple[float, float, float]:
         return (
             (self.upper_bound[0] - self.lower_bound[0]) / self.number_of_cells[0],
             (self.upper_bound[1] - self.lower_bound[1]) / self.number_of_cells[1],
